requirements.txt
asgiref==3.5.0
Django==4.0.3
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
